requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
aiohttp==3.9.0
python-dotenv==1.0.0
slowapi==0.1.8
gunicorn==21.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 